+++ src/components/LineChart/Chart.tsx
@@ -9,6 +9,7 @@ import {
   FONT_SIZE,
   SPACING_TIGHT,
   Margin,
+  OVERFLOW_GRID_RIGHT_MARGIN,
 } from '../../constants';
 import {VisuallyHiddenRows} from '../VisuallyHiddenRows';
 import {LinearXAxis} from '../LinearXAxis';
@@ -25,6 +26,7 @@ import {Crosshair} from '../Crosshair';
 import {LinearGradient} from '../LinearGradient';
 import {ActiveTooltip} from '../../types';
 import {TooltipContainer} from '../TooltipContainer';
+import {HorizontalGridLines} from '../HorizontalGridLines';
 
 import {MAX_ANIMATED_SERIES_LENGTH} from './constants';
 import {
@@ -158,8 +160,12 @@ export function Chart({
 
   const reversedSeries = useMemo(() => series.slice().reverse(), [series]);
 
+  const rightMargin = gridOptions.horizontalOverflow
+    ? OVERFLOW_GRID_RIGHT_MARGIN
+    : Margin.Right;
+
   const drawableWidth =
-    axisMargin == null ? null : dimensions.width - Margin.Right - axisMargin;
+    axisMargin == null ? null : dimensions.width - rightMargin - axisMargin;
 
   const longestSeriesIndex = useMemo(
     () =>
@@ -315,13 +321,24 @@ export function Chart({
         <g transform={`translate(${axisMargin},${Margin.Top})`}>
           <YAxis
             ticks={ticks}
-            drawableWidth={drawableWidth}
             fontSize={fontSize}
-            showGridLines={gridOptions.showHorizontalLines}
-            gridColor={gridOptions.color}
             labelColor={yAxisOptions.labelColor}
           />
         </g>
+
+        {gridOptions.showHorizontalLines ? (
+          <HorizontalGridLines
+            ticks={ticks}
+            color={gridOptions.color}
+            transform={{
+              x: gridOptions.horizontalOverflow ? 0 : axisMargin,
+              y: Margin.Top,
+            }}
+            width={
+              gridOptions.horizontalOverflow ? dimensions.width : drawableWidth
+            }
+          />
+        ) : null}
 
         {emptyState ? null : (
           <g transform={`translate(${axisMargin},${Margin.Top})`}>