+++ src/components/MultiSeriesBarChart/Chart.tsx
@@ -4,6 +4,7 @@ import {TooltipContainer} from '../TooltipContainer';
 import {eventPoint, getTextWidth, getBarXAxisDetails} from '../../utilities';
 import {YAxis} from '../YAxis';
 import {BarChartXAxis} from '../BarChartXAxis';
+import {HorizontalGridLines} from '../HorizontalGridLines';
 
 import {getStackedValues, formatAriaLabel} from './utilities';
 import {
@@ -234,17 +235,30 @@ export function Chart({
           />
         </g>
 
+        {gridOptions.showHorizontalLines ? (
+          <HorizontalGridLines
+            ticks={ticks}
+            color={gridOptions.color}
+            transform={{
+              x: gridOptions.horizontalOverflow ? 0 : axisMargin,
+              y: MARGIN.Top,
+            }}
+            width={
+              gridOptions.horizontalOverflow
+                ? chartDimensions.width
+                : drawableWidth
+            }
+          />
+        ) : null}
+
         <g
           transform={`translate(${axisMargin},${MARGIN.Top})`}
           aria-hidden="true"
         >
           <YAxis
             ticks={ticks}
-            drawableWidth={drawableWidth}
             fontSize={fontSize}
             labelColor={yAxisOptions.labelColor}
-            showGridLines={gridOptions.showHorizontalLines}
-            gridColor={gridOptions.color}
           />
         </g>