+++ src/constants.ts
@@ -26,6 +26,7 @@ export const ROUNDED_BAR_RADIUS = 3;
 export const MIN_BAR_HEIGHT = 5;
 export const EMPTY_STATE_CHART_MIN = 0;
 export const EMPTY_STATE_CHART_MAX = 10;
+export const OVERFLOW_GRID_RIGHT_MARGIN = 20;
 
 export const DEFAULT_GREY_LABEL = 'rgb(99, 115, 129)';
 export const DEFAULT_CROSSHAIR_COLOR = 'rgb(223, 227, 232)';